requirements.txt
pandas
playwright
loguru
openpyxl
elasticsearch<9.0.0
baidu-aip
pillow
requests
chardet
pytesseract
opencv-python